+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next Up
+
+- Improvements to attack and damage parsing roll enrichers in monster features.
+
 # 6.0.87
 
 - [REQUIRES D&D 4.3.5] Updates to DDB Importer to handle changes in the 4.3.x branch of D&D